+++ components/PostOptionsSkeleton.tsx
@@ -1,0 +1,8 @@
+export default function PostOptionsSkeleton() {
+  return (
+    <>
+      <div className="bg-default-neutral flex-grow h-8 skeleton-loading"></div>
+      <div className="bg-default-neutral flex-grow h-8 skeleton-loading"></div>
+    </>
+  );
+}